+++ CHANGES.rst
@@ -7,7 +7,8 @@ UNRELEASED
 
 * Fix ``setup.py`` ``long_description``
 * Allow models fields be defined with class attributes without typing
-* Fix conversion of fields using typing.List[...]
+* Fix type conversion on fields using ``typing.List[...]``
+* Bugfix: remove ``Meta`` attribute from model class meta fields
 
 
 1.0.2 / 2018-01-10